Move MenuActivity class location
