+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         <activity android:name=".admin.DeleteCounter" />
         <activity android:name=".admin.UpdateCounterPass" />
         <activity android:name=".counter.PenjualActivity" />
-        <activity android:name=".counter.Entity.MenuActivity2" />
+        <activity android:name=".counter.MenuActivity" />
         <activity android:name=".counter.EditMenu" />
         <activity android:name=".admin.TambahMenuActivity" />
         <activity android:name=".admin.EditListCounterActivity" />
